correction ft_strlcat.c

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: lnoaille <lnoaille@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/27 17:45:05 by lnoaille          #+#    #+#             */
-/*   Updated: 2020/05/05 17:42:17 by lnoaille         ###   ########.fr       */
+/*   Updated: 2020/05/13 01:25:22 by lnoaille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	long unsigned int pos;
+	size_t pos;
 
 	pos = 0;
 	while (s[pos])
